+++ src/components/Tab/TabList.svelte
@@ -1,0 +1,11 @@
+<div class="tab-list">
+  <slot />
+</div>
+
+<style>
+  .tab-list {
+    /* border-bottom: 1px solid blue; */
+    display: flex;
+    justify-content: space-evenly;
+  }
+</style>