+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 English | [简体中文](./CHANGELOG_CN.md)
 
+#### 3.9.0 (2021-07-16)
+
+- `Feat(Log)` Show audio loading error log. (PR #419 by @zimv)
+- `Feat(Storage)` Rewrite Storage panel, supports add/edit/delete storage items. (PR #429 by @ManiaciaChao)
+- `Feat(Plugin)` New third-party plugin [vite-plugin-vconsole](https://github.com/vadxq/vite-plugin-vconsole). (by @vadxq)
+- `Refactor(Global)` Use Svelte as template engine. (PR #429 by @ManiaciaChao)
+- `Refactor(Core|Element)` Convert core file and Element panel to `.ts` file.
+- `Fix(Log)` Fix error when print object(s) with no `toJSON` method such as `Vue` instance. (PR #431 by @sillyhong)
+- `Fix(Network)` Fix error when url not starts with `http`. (issue #420)
+- `Fix(Network)` Fix error when using `Request` as `Fetch`'s parameter. (PR #428 by @tatsunoneko)
+- `Fix(Network)` Display formatted key-value list when `POST` a JSON string. (issue #425)
+- `style` Wrap LESS math operations. (PR #426 by @ManiaciaChao)
+- `Chore` Fix `yarn serve` error. (issue #424)
+
+
 #### 3.8.1 (2021-06-24)
 
 - `Fix(Log)` Do not merge repeated logs with object(s) or array(s) into one line to avoid merging objects with the same structure but different values.